requirements.txt
# Web Framework
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0

# Image and Video Processing
opencv-python-headless==4.8.0.76
Pillow==9.5.0

# Machine Learning and Numerical Computing
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
git+https://github.com/openai/CLIP.git

# HTTP Requests
requests==2.31.0

# Utils
python-dotenv==1.0.0